hello_world/trash/backend/requirements.txt
fastapi
uvicorn
python-arango
pydantic
python-multipart